Do not set file attributes during read operations
For read-only files, it certainly does not make sense to try to overwrite attributes.
For read-write access, it probably does not make sense, either.
For newly-created read-write files, it would make sense but the create operation sets
the read-write flag by default, so this is redundant.

So I am removing these calls.
This fixes #1.